test(tabs): migrate placements
