+++ core/src/components/tabs/test/placements/tabs.e2e.ts-snapshots/tabs-tab-bar-bottom-md-ltr-Mobile-Chrome-linux.png
