+++ core/src/components/tabs/test/placements/tabs.e2e.ts-snapshots/tabs-tab-bar-top-md-ltr-Mobile-Firefox-linux.png
